# [ArrowStringArray] PERF: Series.str.get_dummies