planning to remove the padding code from _wrap_result eventually, but until then we can skip it when we return a integer array from get_dummies

adding tests and benchmarks as precursor to potential changes to _wrap_result  #41372

have a working implementation for ArrowStringArray using pyarrow native functions but is slower than object fallback, so am leaving that for a followup.

```
       before           after         ratio
     [4ec6925c]       [091b0b02]
     <master>         <get_dummies>
-      2.58±0.02s         655±10ms     0.25  strings.Dummies.time_get_dummies('arrow_string')
-      2.58±0.03s          643±9ms     0.25  strings.Dummies.time_get_dummies('string')
-      2.59±0.07s          638±7ms     0.25  strings.Dummies.time_get_dummies('str')

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
PERFORMANCE INCREASED.
```